fix: cast water_ml from JSONB as numeric not int
_clean_micros() stores water_ml as a float (e.g. 500.0), so
micros->>'water_ml' returns '500.0'. Casting '500.0'::int raises a
PostgreSQL error; ::numeric handles both integer and float strings.

+++ backend/app/nutrition.py
@@ -271,7 +271,7 @@ def history(chat_id: str, days: int = 14) -> list[dict]:
                        SUM(carbs)    FILTER (WHERE source != 'water') AS carbs,
                        SUM(calories) FILTER (WHERE source != 'water') AS calories,
                        COUNT(*)      FILTER (WHERE source != 'water') AS meals,
-                       SUM(COALESCE((micros->>'water_ml')::int, 0))
+                       SUM(COALESCE((micros->>'water_ml')::numeric, 0))
                          FILTER (WHERE source = 'water')              AS water_ml
                 FROM nutrition_logs
                 WHERE chat_id = %s AND log_date >= %s
